fix: localDate 파싱 변경 및 purge response 추가

+++ backend/src/main/java/org/hikikomori/community/repository/CommentRepository.java
@@ -14,5 +14,5 @@ public interface CommentRepository extends JpaRepository<Comment, UUID> {
 
     @Modifying
     @Query("DELETE FROM Comment c WHERE c.createdAt >= :startAt AND c.createdAt < :endAt")
-    int deleteByCreatedAtBetween(LocalDateTime startAt, LocalDateTime endAt);
+    long deleteByCreatedAtBetween(LocalDateTime startAt, LocalDateTime endAt);
 }
